feat: Ajouter les règles de validation du mot de passe lors de la création de compte
