+++ app/lib/features/authentification/creer_compte/presentation/pages/creer_compte_view.dart
@@ -3,6 +3,7 @@ import 'package:app/features/authentification/creer_compte/presentation/blocs/cr
 import 'package:app/features/authentification/creer_compte/presentation/blocs/creer_compte_state.dart';
 import 'package:app/features/authentification/saisie_code/presentation/pages/saisie_code_page.dart';
 import 'package:app/l10n/l10n.dart';
+import 'package:app/shared/widgets/composants/mot_de_passe/mot_de_passe.dart';
 import 'package:dsfr/dsfr.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -17,10 +18,6 @@ class CreerCompteView extends StatelessWidget {
 
   void _handleMotDePasse(final BuildContext context, final String value) {
     context.read<CreerCompteBloc>().add(CreerCompteMotDePasseAChange(value));
-  }
-
-  void _handleCreerCompte(final BuildContext context) {
-    context.read<CreerCompteBloc>().add(const CreerCompteCreationDemandee());
   }
 
   Future<void> _handleCompteCree(
@@ -58,20 +55,30 @@ class CreerCompteView extends StatelessWidget {
               keyboardType: TextInputType.emailAddress,
             ),
             const SizedBox(height: DsfrSpacings.s2w),
-            DsfrInput(
-              label: Localisation.motDePasse,
+            FnvMotDePasse(
               onChanged: (final value) => _handleMotDePasse(context, value),
-              isPasswordMode: true,
-              keyboardType: TextInputType.visiblePassword,
             ),
             const SizedBox(height: DsfrSpacings.s2w),
-            DsfrButton(
-              label: Localisation.creerMonCompte,
-              variant: DsfrButtonVariant.primary,
-              size: DsfrButtonSize.lg,
-              onPressed: () => _handleCreerCompte(context),
-            ),
+            const _BoutonCreerCompte(),
           ],
         ),
+      );
+}
+
+class _BoutonCreerCompte extends StatelessWidget {
+  const _BoutonCreerCompte();
+
+  @override
+  Widget build(final BuildContext context) => DsfrButton(
+        label: Localisation.creerMonCompte,
+        variant: DsfrButtonVariant.primary,
+        size: DsfrButtonSize.lg,
+        onPressed: context.select<CreerCompteBloc, bool>(
+          (final bloc) => bloc.state.estValide,
+        )
+            ? () => context
+                .read<CreerCompteBloc>()
+                .add(const CreerCompteCreationDemandee())
+            : null,
       );
 }
