+++ app/lib/features/profil/changer_mot_de_passe/presentation/widgets/changer_mot_de_passe.dart
@@ -3,6 +3,7 @@ import 'package:app/features/profil/changer_mot_de_passe/presentation/blocs/chan
 import 'package:app/features/profil/changer_mot_de_passe/presentation/blocs/changer_mot_de_passe_state.dart';
 import 'package:app/features/profil/presentation/widgets/profil_titre_partie.dart';
 import 'package:app/l10n/l10n.dart';
+import 'package:app/shared/widgets/composants/mot_de_passe/mot_de_passe.dart';
 import 'package:dsfr/dsfr.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -52,6 +53,17 @@ class _ChangerMotDePasse extends StatelessWidget {
       );
 }
 
+class _MotDePasse extends StatelessWidget {
+  const _MotDePasse();
+
+  @override
+  Widget build(final BuildContext context) => FnvMotDePasse(
+        onChanged: (final value) => context
+            .read<ChangerMotDePasseBloc>()
+            .add(ChangerMotDePasseAChange(value)),
+      );
+}
+
 class _ChangerMotDePasseButton extends StatelessWidget {
   const _ChangerMotDePasseButton();
 
@@ -67,93 +79,5 @@ class _ChangerMotDePasseButton extends StatelessWidget {
                 .read<ChangerMotDePasseBloc>()
                 .add(const ChangerMotDePasseChangementDemande())
             : null,
-      );
-}
-
-class _MotDePasse extends StatelessWidget {
-  const _MotDePasse();
-
-  @override
-  Widget build(final BuildContext context) => Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          DsfrInput(
-            label: Localisation.motDePasse,
-            onChanged: (final value) => context
-                .read<ChangerMotDePasseBloc>()
-                .add(ChangerMotDePasseAChange(value)),
-            isPasswordMode: true,
-            keyboardType: TextInputType.visiblePassword,
-          ),
-          const Text(
-            Localisation.votreMotDePasseDoitContenir,
-            style: DsfrTextStyle.bodyXs(color: DsfrColors.grey425),
-          ),
-          const _DouzeCarateresMinimum(),
-          const _AuMoinsUnMajusculeEtUneMinuscule(),
-          const _UnCaractereSpecialMinimum(),
-          const _UnChiffreMinimum(),
-        ],
-      );
-}
-
-class _DouzeCarateresMinimum extends StatelessWidget {
-  const _DouzeCarateresMinimum();
-
-  @override
-  Widget build(final BuildContext context) => _DoitContenir(
-        valid: context.select<ChangerMotDePasseBloc, bool>(
-          (final bloc) => bloc.state.douzeCaracteresMinimum,
-        ),
-        text: Localisation.motDePasse12CaractresMinimum,
-      );
-}
-
-class _AuMoinsUnMajusculeEtUneMinuscule extends StatelessWidget {
-  const _AuMoinsUnMajusculeEtUneMinuscule();
-
-  @override
-  Widget build(final BuildContext context) => _DoitContenir(
-        valid: context.select<ChangerMotDePasseBloc, bool>(
-          (final bloc) => bloc.state.auMoinsUneMajusculeEtUneMinuscule,
-        ),
-        text: Localisation.motDePasse1MajusculeEt1Minuscule,
-      );
-}
-
-class _UnCaractereSpecialMinimum extends StatelessWidget {
-  const _UnCaractereSpecialMinimum();
-
-  @override
-  Widget build(final BuildContext context) => _DoitContenir(
-        valid: context.select<ChangerMotDePasseBloc, bool>(
-          (final bloc) => bloc.state.unCaractereSpecialMinimum,
-        ),
-        text: Localisation.motDePasse1CaractreSpecialMinimum,
-      );
-}
-
-class _UnChiffreMinimum extends StatelessWidget {
-  const _UnChiffreMinimum();
-
-  @override
-  Widget build(final BuildContext context) => _DoitContenir(
-        valid: context.select<ChangerMotDePasseBloc, bool>(
-          (final bloc) => bloc.state.unChiffreMinimum,
-        ),
-        text: Localisation.motDePasse1ChiffreMinimum,
-      );
-}
-
-class _DoitContenir extends StatelessWidget {
-  const _DoitContenir({required this.valid, required this.text});
-
-  final bool valid;
-  final String text;
-
-  @override
-  Widget build(final BuildContext context) => DsfrFormMessage(
-        type: valid ? DsfrFormMessageType.valid : DsfrFormMessageType.info,
-        text: text,
       );
 }
